Add initial implementation as a Nix Flake
This is basically the same implementation I did post as a Gist[1] a
while ago, but turned into a Flake, so that I can re-use it from various
projects.

I also made sure that certificates last for 1000 years, which is
hopefully enough to make sure we don't run into cert failures anytime
soon. Ideally, we'd just let it never expire, but I vaguely remember
there being TLS implementations that don't like this.

[1]: https://gist.github.com/aszlig/a836d25a467b354732506cc77499f56a

Signed-off-by: aszlig <aszlig@nix.build>