Use higher serial for non-CA certs
When using the same serial as the CA root certificate, newer Firefox (or
to be more exact: newer NSS) versions will bail out with
SEC_ERROR_REUSED_ISSUER_AND_SERIAL.

To avoid this, let's simply use a higher serial number for all server
certificates by just patching MiniCA again. I know this is truly ugly
and I'm not proud of this, but have I mentioned that you shouldn't use
this for anything but testing? ;-)

Signed-off-by: aszlig <aszlig@nix.build>